tests/unit/test_employee.py
"""
Unit tests for employee module
Tests models, views, forms, and serializers
"""

import os
import sys
import django

# Add project root to Python path
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))))

# Setup Django
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
django.setup()

from django.test import TestCase
from django.urls import reverse
from django.core.exceptions import ValidationError

from .base import TestBase


class EmployeeModelTest(TestBase):
    """Test employee models"""
    
    def test_model_creation(self):
        """Test model creation"""
        # TODO: Implement model creation tests
        pass
        
    def test_model_validation(self):
        """Test model validation"""
        # TODO: Implement model validation tests
        pass


class EmployeeViewsTest(TestBase):
    """Test employee views"""
    
    def test_view_access(self):
        """Test view access"""
        # TODO: Implement view access tests
        pass
        
    def test_view_permissions(self):
        """Test view permissions"""
        # TODO: Implement view permission tests
        pass


class EmployeeFormsTest(TestBase):
    """Test employee forms"""
    
    def test_form_validation(self):
        """Test form validation"""
        # TODO: Implement form validation tests
        pass


class EmployeeSerializersTest(TestBase):
    """Test employee serializers"""
    
    def test_serializer_serialization(self):
        """Test serializer serialization"""
        # TODO: Implement serializer tests
        pass
