tests/requirements.txt
# Test dependencies
watchdog>=3.0.0
coverage>=7.0.0
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0
